+++ R/mirtCAT.R
@@ -98,7 +98,9 @@
 #'    \code{df} object, however it provides a great deal of customization, 
 #'    particularly through the use of \code{div()} and other helpful tags. When used, this list
 #'    must have the same number of elements as rows in the \code{df} object, and items that
-#'    should not use this input should be filled with \code{NULL}
+#'    should not use this input should be filled with \code{NULL}. Alternatively, if specified the
+#'    names of the elements to this list can be used to match the rownames of the \code{df} object
+#'    to avoid the use of NULL placeholders.
 #'       
 #'    E.g., the following would result in a bolded and italicized item stems for the first two items in a 
 #'    three item test: 
@@ -109,7 +111,14 @@
 #'    stacked; for example, \code{div(HTML("This is some HTML"), br(), HTML("And in the middle a line break")} 
 #'    can be used as a single element in the \code{shinyStems} list.
 #'    See \code{http://shiny.rstudio.com/articles/tag-glossary.html} for more examples of how
-#'    to use tags and HTML generating functions.
+#'    to use tags and HTML generating functions. Note that if names are supplied then the following is equivalent
+#'    (i.e., the NULL elements can be dropped)
+#'    
+#'    \code{shinyStems <- list("1"=strong('Stem 1'), "2"=em('Stem 2'))}
+#'    
+#'    assuming that respective items are named "1" and "2" in the \code{df} object.
+#'    
+#'    
 #'    
 #' @param method argument passed to \code{mirt::fscores()} for computing new scores in the CAT 
 #'   stage, with the addition of a \code{'fixed'} input to keep the latent trait estimates
@@ -515,7 +524,10 @@
 #'      the input is 'radio' or 'checkbox' (see \code{\link{radioButtons}}). 
 #'      This is used to modify the output of the controllers using 
 #'      suitable HTML code. If a row in \code{df} should not have a customized names then supplying 
-#'      the value \code{NA} in the associated list element will use the standard inputs instead}
+#'      the value \code{NULL} in the associated list element will use the standard inputs instead. 
+#'      Alternatively, if specified the names of the elements to this list can be used to match the 
+#'      rownames of the \code{df} object to avoid the use of \code{NULL} placeholders. See \code{shinyStems}
+#'      for an example of this behaviour.}
 #'      
 #'    \item{\code{choiceValues}}{associated values to be used along with \code{choiceNames} (see above)}
 #'